Fix inclusion of config files in source distribution
The example config files were not included in the source distribution
uploaded to pypi. Reviewing the relevant documentation [1] [2] showed
that the fix is fairly simple - the files must be listed in MANIFEST.in
and the syntax used in package_data in setup.py was wrong. We don't
actually need to set package_data in setup.py once the files are listed
in MANIFEST.in since we already have include_package_data=True.

[1]: https://packaging.python.org/guides/using-manifest-in/
[2]: https://setuptools.readthedocs.io/en/latest/userguide/datafiles.html