Remove duplicate config files under for_build
Instead of duplicating the config files we can use a symlink. This
reduces confusion and avoids the risk of the two config copies diverging
over time.

When installing these files in `make_packages.sh` and
`generate_deb_package.sh`, we now use the `-L` argument to cp to
dereference the link.